Diretório 'data' criado e suas subpastas 'card' e 'form' contendo seus respectivos arquivos com dados. Diretório 'utils' criado com a subpasta 'date', para funções auxiliares. Componentes 'CardList' e 'Form' e page 'LicitacaPage' refatorados
