+++ src/components/form/Form.tsx
@@ -1,33 +1,19 @@
 import { useState } from 'react';
-import './Form.css';
 import { FaAngleRight, FaAngleDown } from 'react-icons/fa';
-
+import './Form.css';
+import { modalidadeOptions, situacaoOptions, unidadeGestoraOptions } from '../../data/form/formData';
+import { formatDate, nextYear, today } from '../../utils/date/FormatDate';
 
 interface FormData {
   unidadeGestora: string;
-  dataInicio: string;
-  dataFim: string;
-  modalidade: string;
-  situacao: string;
-  palavraChave: string;
+  dataInicio?: string;
+  dataFim?: string;
+  modalidade?: string;
+  situacao?: string;
+  palavraChave?: string;
 }
 
-const unidadeGestoraOptions = ['IFSC', 'Prefeitura', 'Outros'];
-const modalidadeOptions = ['Pregão', 'Concorrência', 'Dispensa'];
-const situacaoOptions = ['Aberta', 'Encerrada', 'Em andamento'];
-
-const today = new Date();
-const nextYear = new Date();
-nextYear.setFullYear(today.getFullYear() + 1);
-
-const formatDate = (date: Date) => {
-  const day = String(date.getDate()).padStart(2, '0');
-  const month = String(date.getMonth() + 1).padStart(2, '0');
-  const year = date.getFullYear();
-  return `${day}/${month}/${year}`;
-};
-
-const Form: React.FC = () => {
+function Form() {
   const [formData, setFormData] = useState<FormData>({
     unidadeGestora: '',
     dataInicio: formatDate(today),
@@ -91,7 +77,6 @@ const Form: React.FC = () => {
               : ''
           }
           onChange={handlePeriodoChange}
-          required
         />
       </div>
 
@@ -103,7 +88,6 @@ const Form: React.FC = () => {
             name="modalidade"
             value={formData.modalidade}
             onChange={handleInputChange}
-            required
           >
             <option value="">Selecione</option>
             {modalidadeOptions.map(opt => (
@@ -122,7 +106,6 @@ const Form: React.FC = () => {
             name="situacao"
             value={formData.situacao}
             onChange={handleInputChange}
-            required
           >
             <option value="">Selecione</option>
             {situacaoOptions.map(opt => (
@@ -141,7 +124,6 @@ const Form: React.FC = () => {
           name="palavraChave"
           value={formData.palavraChave}
           onChange={handleInputChange}
-          required
         />
       </div>
