+++ src/data/form/formData.ts
@@ -1,0 +1,3 @@
+export const unidadeGestoraOptions = ['IFSC', 'Prefeitura', 'Outros'];
+export const modalidadeOptions = ['Pregão', 'Concorrência', 'Dispensa'];
+export const situacaoOptions = ['Aberta', 'Encerrada', 'Em andamento'];
