+++ src/utils/date/FormatDate.ts
@@ -1,0 +1,10 @@
+export const today = new Date();
+export const nextYear = new Date();
+nextYear.setFullYear(today.getFullYear() + 1);
+
+export const formatDate = (date: Date) => {
+  const day = String(date.getDate()).padStart(2, '0');
+  const month = String(date.getMonth() + 1).padStart(2, '0');
+  const year = date.getFullYear();
+  return `${day}/${month}/${year}`;
+};
